Revision 18
To do
-----
1. How about checking frequency manually one channel by one channel

Prologue
--------
[02-Oct-16 11:45AM]
1. Revision 17 seems ok with a few blocks(30 min interval) not sending.
Still cannot pinpoint what channel is not sending though. The gateway
hang at about midnight
2. The other point that I want to point out is I missed out on RX2
frequency which is 867.2Mhz so it the channel should be

Channel: 6,7,8,9,10,RX2
Freq: 866.2,866.4,866.6,866.8,867,867.2

3. Oh another point is that revision 17 is dealing with channel 6-10 so
i don't know if it is going to cause any problem

Update
-------
[02 Oct 16 03:32PM]
Change sleeptime to 5 minute to test out determine working channel

[02-Oct-16 07:39PM]
Change the data rate from 0-5 to 3-5. The reason for this is I want is
to reduce the time on air and to test why the hell does it keep sending
on SF12.
Answer: When i changed it data rate to 3-5, suddenly all tranmission
stopped. I am guessing it is because of the radio power is weak

[02-Oct-16 08:22PM]
Change data rate from 3-5 to 1-5. I suspect it only work in SF12.

[02-Oct-16 11:18PM]

I used another an old version to set the the data rate to SF7 and it
works. I think the line of code that did it is this:

error = LoRaWAN.setDataRate(DR);

One thing to note is that once the configuration is saved, it will
stick. For example if I already set channel 3 to become 866.2 then it
will always be 866.2 until you turn it off or something like that. I
will need to do more confirmation on this.

One more thing i noticed is that what I tried to do earlier is to not
use channel 3-5 and use 6-10. However, RN2483 don't allow you to do that
instead it will just make whatever is the channel that is not used to be
the starting channel. For my case channel 6 become channel 3 and channel
7 become channel 4 till channel 10 become channel 7. How do i know
because he channel correspond to a specific frequency that I set. So it
looks something like the following now.

Channel: 3,4,5,6,7,RX2
Freq: 866.2,866.4,866.6,866.8,867,867.2

[03-Oct-16 02:46PM]
Set adaptive data rate to ON

[03-Oct-16 02:46PM]
After consolidation of the various frequency not using adaptive data
rate, conduit seems head into the direction of SF12. In the end it just
became all SF12

[03-Oct-16 04:51PM]
After setting the adaptive data rate to ON for some reason Channel 0 to
2 seems to be turned on for some reason. It uses illegal channel of
868.1 Mhz, 868.3 Mhz and 868.5 Mhz.

[03-Oct-16 05:00PM]
Lets try to enable channel 3 865Mhz to channel 5 865.4 Mhz

[03-Oct-16 05:32PM]
Try moving adaptive data rate on to after setting RX2 parameters. This
means switch off channel 0-2 first and then turn on. Instead of turn on
first and then turn off

[03-Oct-16 05:58PM]
ADR keep persisting to use channel 0-2. I have no choice but to turn it
off and configure the channel to use configure data rate 2 to data rate
5