Revision 19
[03-Oct-16 07:07PM]
Finally I am removing channel 865, 865.2 and 865.4 and replace
officially setting channel 3 to 866.2 and so on and so forth.

Also I must turn off channel 8 to 10

[03-Oct-16 08:18PM]
Turn on ADR again. The good thing about ADR is that (from my record at
least), it will connect and tranmit the payload to the gateway each
time. The bad thing is that it will use channel 0 to 2 even though it
has been explicitly turned off.

[03-Oct-16 11:16PM]
1. Try using radio power of 2 dbm. It works well given such a short
distance between gateway and mote
2. Now change to 14 dbm & turn off ADR

[04-Oct-16 12:25PM]
After all these setting it turns out that Plug & Sense is sending at SF7
or SF8 most of the time. In fact, SF7 and SF8 are the data rate that is
always been sending throughout the legal channel i.e channel 3 to 7. I
have connected everything back to M2X and everything is working again.

[04-Oct-16 12:31PM]
Now I want to test the M2X node that comes preinstalled with the new
firmware and I want to see whether it will crash node red again or not.
But this time I want to send the data to virtual device in M2X. Let's do
it. Apparently, it still crash.

[04-Oct-16 12:51PM]
Set sleep time to 10 min in order to follow a duty cycle of <1%. In fact
15 min is very long. A device transmitting 0.5s long need need to be
unavailable for 49.5s long in order to follow a duty cycle of 1%.
Therefore me setting it to 10 is more than gracious enough.

[04-Oct-16 04:13PM]
For some reason, conduit initialized itself back to factory setting. I
have to go through the hoop of configuring everything again. But, it is
not that hard.

[04-Oct-16 05:13PM]
Clear LoRa history

[04-Oct-16 07:32PM]
Update M2X IoT Cloud Service Flow