Revision 20
[04-Oct-16 11:15PM]
Everything works fine. It is just that I still could not figure out why
sometimes the RN2483 in Plug & sense just wouldn't want to send a
mac_tx_ok and sometimes it wants to send it. Is it dependant on the data
rate? I suspect it is.

I play around again with the RX1Delay. Right now i set the RX1 delay to
0 sec

[04-Oct-16 11:30PM]
Set the DR from SF7 to SF9 and see how it goes
Answer: still the same

[05-Oct-16 12:10AM]
Change LoRa network server Tx Power to 10dbm

[06-Oct-16 12:53PM]
I checked that M2X is not receiving any data. My previous suspect is
that node-red hang. However today, I found out that node-red did not
hang and I am able to still inject manual data to M2X to another device
which I called "Virtual Device". I am positive it is not node red
problem.

So the problem might be the LoRa Network Server or it is RN2483 not
transmitting. The reason I suspect it is lora-network-server is because
I cannot see the LoRa statistic.

[06-Oct-16 2:35PM]
Turn on and off plug and sense but it seems that the lora-network-server
is not responding. For one thing I could not see the statistic and
secondly nothing is appearing in node red. So I know the network server
is down.

[06-Oct-16 2:43PM]
Restart Lora network server. I tested 2 method one of them is restart
which give me a warning of "failed to kill 1119: No such process". After
that I switched off and switched on plug & sense but I still could not
get a read.

The second method that I used is to issue stop command to the
lora-network-server. This time it really stopped. After I confirmed that
it has stopped without error only then do I start the
lora-network-server. Everything goes great. Conduit start receiving the
data and my node-red is responding.

Conclusion: I think I am quite sure that it is the lora-network-server
that is causing the problem. Why does the server hung on itself? that is
a good question. I think it is because of the number of CRC error and
duplicates join and duplicate ack that is causing the problem. Actually,
I think the real reason is because firmware 1.3.2 is not stable.

[06-Oct-16 2:43PM]
Updated node-red flow to include battery information to M2X

[06-Oct-16 4:43PM]
I downgrade the conduit firmware to 1.2.2. Lo and behold, this version
is much more stable. Even though the apps page in the web server is
giving me a bit of problem and it mentioned that node-red is not
enabled. However it is enabled. The other thing is that SSH is also
running smoothly in firmware 1.2.2. The whole web server is also very
smooth in 1.2.2.

[06-Oct-16 7:00PM]
Modified flows which add control flow inside M2X IoT Cloud Service flow.
The control flow will allow me to test whether node-red is down or not.
If node-red is not working then even control flow won't execute and send
data to M2X