Revision 21
[07-Oct-16 02:00PM]
Change to 30 minute sleeptime as apparently there is spam attacking. I
don't know how.
Ans: After getting really serious about looking into this problem. I
have finally figured it out. It all boils down to 1 problem. That is the
port to the conduit is open from the router side. I have tried to change
password but it is all futile. The hacker can just use the password that
you have key-ed in previously from HTTP and somehow use it. The most
that I got is from Guangzhou. The hacker is really good.

[07-Oct-16 04:00PM]
Configure firewall for conduit. I have configured inbound and outbound
port of 22, 80 & 443. At first I did not include port 80 and http
request could not be performed in node-red. That is when I know i need
to enable port 80. After enabling port 80. I can send data to M2X using
the control flow.

[07-Oct-16 06:00PM]
I guess I want to test out again whether I will be suspended again since
now I have disabled all the port from sending except these 3 ports.
Let's set the sleep time to 20min and change some of the channel to
incorportate minimum data rate of 7 or 8 because apparently, plug and
sense keep on sending SF9. I set channel 3 data rate to SF7 only,
channel 4 SF7 to SF8, Channel 5 SF7 to SF9, channel 6 SF7 only, channel
7 SF7 to SF8.
Ans: Used in conjuction with lora-network-server - 1.0.8 r0.0 it seems
to be only want to transmit at the highest SF that is configured. In
this case it is only transferring at SF9 and if it is lower then it
refused to transmit.

[08-Oct-16 04:13PM]
Changed password to a 20 character. That should be sufficient not to get
hacked!

[08-Oct-16 10:50PM]
Change parameter in Lora Network Server Configuration > Network Server
Logging
Log Destination: File
Path: /var/log/lora-network-server.log
Log Level: Maximum

[11-Oct-16 12:48PM]
After playing around with learning linux especially with the CLI and how
to execute system process like CPU usage, disk usage and most
importantly how to see that a hacker has got into your system. It is
because of this piece of informatio that I know that the hacker is
always targeting on my port 22 and at that time my port 22 is always
open and after closing the port, I am very careful in always checking
that nobody messes with my system. After that all the system is smooth.
Smooth in terms of web GUI, the gateway does not hang, and the gateway
is constantly sending data to M2X.

Now what I want to change is I want to change the data rate to be the
same (again) that means that all the channel will be tranmitting from
SF7 to SF9 irregardless. This is because on the 07th october at 6pm i
tried using different channel to different data rate and it turns out
that some channel is not transmitting. Let's see how it goes this time.